+++ tkclasswiz/__init__.py
@@ -26,7 +26,9 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
-__version__ = "1.3.1"
+
+__version__ = "1.4.0"
+
 
 from .object_frame import *
 from .annotations import *
@@ -37,3 +39,4 @@ from .messagebox import *
 from .storage import *
 from .utilities import *
 from .aliasing import *
+from .deprecation import *