+++ tkclasswiz/object_frame/frame_base.py
@@ -1,4 +1,5 @@
 from typing import get_args, get_origin, Iterable, Union, Literal, Any, TYPE_CHECKING, TypeVar, Generic
+from abc import ABC, abstractmethod
 from inspect import isabstract
 from contextlib import suppress
 from itertools import chain
@@ -32,7 +33,7 @@ __all__ = (
 
 @extendable
 @doc_category("Object frames")
-class NewObjectFrameBase(ttk.Frame):
+class NewObjectFrameBase(ttk.Frame, ABC):
     """
     Base Frame for inside the :class:`ObjectEditWindow` that allows object definition.
 
@@ -122,6 +123,36 @@ class NewObjectFrameBase(ttk.Frame):
         cls.origin_window = window
 
     @classmethod
+    def filter_literals(cls, types_: Iterable):
+        "Returns only the literal types from ``types_``"
+        return [x for x in types_ if get_origin(x) is Literal]
+
+    @classmethod
+    def check_literals(cls, value: str, literal_types: Iterable):
+        """
+        Checks if the ``value`` is a correct iterable.
+        
+        Parameters
+        -------------
+        value: str
+            The string value to be checked.
+        types: Iterable[Literal]
+            An iterable of accepted literals.
+        """
+        allowed_values = []
+        for literal_type in literal_types:
+            args = get_args(literal_type)
+            if value in args:
+                return value
+
+            allowed_values.extend(args)
+
+        raise ValueError(
+            f"'{value}' (a string) is not one of accepted types and does not match any literal values.\n"
+            f"'Allowed literals: {allowed_values}."
+        )
+
+    @classmethod
     def cast_type(cls, value: Any, types: Iterable):
         """
         Tries to convert *value* into one of the types inside *types* (first successful).
@@ -135,13 +166,6 @@ class NewObjectFrameBase(ttk.Frame):
         CAST_FUNTIONS = {
             dict: lambda v: convert_to_object_info(json.loads(v))
         }
-
-        # Validate literals
-        if get_origin(types[0]) is Literal:
-            if value not in (args := get_args(types[0])):
-                raise ValueError(f"'{value}' is not a valid value'. Accepted: {args}")
-            
-            return value
 
         for type_ in filter(lambda t: t.__module__ == "builtins", types):
             with suppress(Exception):
@@ -264,12 +288,14 @@ class NewObjectFrameBase(ttk.Frame):
             class_, widget, allow_save=allow_save, *args, **kwargs
         )
 
+    @abstractmethod
     def to_object(self):
         """
         Creates an object from the GUI data.
         """
         raise NotImplementedError
 
+    @abstractmethod
     def load(self, old_data: Any):
         """
         Loads the old object info data into the GUI.
@@ -305,6 +331,7 @@ class NewObjectFrameBase(ttk.Frame):
         """
         self._original_gui_data = self.get_gui_data()
 
+    @abstractmethod
     def get_gui_data(self) -> Any:
         """
         Returns all GUI values.